requirements.txt
Markdown==3.10.2
Pygments==2.19.2
PyYAML==6.0.3
watchdog==6.0.0
pytest==9.0.2
